Radar auth integration

+++ src/main/java/org/radarcns/management/web/rest/SensorDataResource.java
@@ -1,24 +1,36 @@
 package org.radarcns.management.web.rest;
 
 import com.codahale.metrics.annotation.Timed;
-import org.radarcns.management.security.AuthoritiesConstants;
-import org.radarcns.management.service.SensorDataService;
-import org.radarcns.management.web.rest.util.HeaderUtil;
-import org.radarcns.management.service.dto.SensorDataDTO;
 import io.github.jhipster.web.util.ResponseUtil;
+import org.radarcns.management.service.SensorDataService;
+import org.radarcns.management.service.dto.SensorDataDTO;
+import org.radarcns.management.web.rest.util.HeaderUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.annotation.Secured;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import java.net.URI;
 import java.net.URISyntaxException;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
+
+import static org.radarcns.auth.authorization.Permission.SENSORDATA_CREATE;
+import static org.radarcns.auth.authorization.Permission.SENSORDATA_DELETE;
+import static org.radarcns.auth.authorization.Permission.SENSORDATA_READ;
+import static org.radarcns.auth.authorization.Permission.SENSORDATA_UPDATE;
+import static org.radarcns.auth.authorization.RadarAuthorization.checkPermission;
+import static org.radarcns.management.security.SecurityUtils.getJWT;
 
 /**
  * REST controller for managing SensorData.
@@ -31,11 +43,11 @@ public class SensorDataResource {
 
     private static final String ENTITY_NAME = "sensorData";
 
-    private final SensorDataService sensorDataService;
+    @Autowired
+    private SensorDataService sensorDataService;
 
-    public SensorDataResource(SensorDataService sensorDataService) {
-        this.sensorDataService = sensorDataService;
-    }
+    @Autowired
+    private HttpServletRequest servletRequest;
 
     /**
      * POST  /sensor-data : Create a new sensorData.
@@ -46,9 +58,9 @@ public class SensorDataResource {
      */
     @PostMapping("/sensor-data")
     @Timed
-    @Secured({AuthoritiesConstants.SYS_ADMIN, AuthoritiesConstants.PROJECT_ADMIN})
     public ResponseEntity<SensorDataDTO> createSensorData(@Valid @RequestBody SensorDataDTO sensorDataDTO) throws URISyntaxException {
         log.debug("REST request to save SensorData : {}", sensorDataDTO);
+        checkPermission(getJWT(servletRequest), SENSORDATA_CREATE);
         if (sensorDataDTO.getId() != null) {
             return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert(ENTITY_NAME, "idexists", "A new sensorData cannot already have an ID")).body(null);
         }
@@ -69,12 +81,12 @@ public class SensorDataResource {
      */
     @PutMapping("/sensor-data")
     @Timed
-    @Secured({AuthoritiesConstants.SYS_ADMIN, AuthoritiesConstants.PROJECT_ADMIN})
     public ResponseEntity<SensorDataDTO> updateSensorData(@Valid @RequestBody SensorDataDTO sensorDataDTO) throws URISyntaxException {
         log.debug("REST request to update SensorData : {}", sensorDataDTO);
         if (sensorDataDTO.getId() == null) {
             return createSensorData(sensorDataDTO);
         }
+        checkPermission(getJWT(servletRequest), SENSORDATA_UPDATE);
         SensorDataDTO result = sensorDataService.save(sensorDataDTO);
         return ResponseEntity.ok()
             .headers(HeaderUtil.createEntityUpdateAlert(ENTITY_NAME, sensorDataDTO.getId().toString()))
@@ -90,6 +102,7 @@ public class SensorDataResource {
     @Timed
     public List<SensorDataDTO> getAllSensorData() {
         log.debug("REST request to get all SensorData");
+        checkPermission(getJWT(servletRequest), SENSORDATA_READ);
         return sensorDataService.findAll();
     }
 
@@ -103,6 +116,7 @@ public class SensorDataResource {
     @Timed
     public ResponseEntity<SensorDataDTO> getSensorData(@PathVariable Long id) {
         log.debug("REST request to get SensorData : {}", id);
+        checkPermission(getJWT(servletRequest), SENSORDATA_CREATE);
         SensorDataDTO sensorDataDTO = sensorDataService.findOne(id);
         return ResponseUtil.wrapOrNotFound(Optional.ofNullable(sensorDataDTO));
     }
@@ -117,6 +131,7 @@ public class SensorDataResource {
     @Timed
     public ResponseEntity<Void> deleteSensorData(@PathVariable Long id) {
         log.debug("REST request to delete SensorData : {}", id);
+        checkPermission(getJWT(servletRequest), SENSORDATA_DELETE);
         sensorDataService.delete(id);
         return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert(ENTITY_NAME, id.toString())).build();
     }
